Moved status connection inside ping
Moved status connection attempt inside of ping. In the case of a connection failure, embark would have to complete a build in order for a reconnection attempt to occur. Now, reconnection attempts happen on every ping, until a connection is successful.

Handles error messages from status app and logs them.