Refactor of library
Much of the index was rewritten to get a final workflow in place.

The plugin will now attempt to connect to an already added network first before adding a new network to the app. Once connected to the network, the dapp will open using the IP of the local machine.

There is more work to come regarded moving to ES6 with a babel transpilation.