fix(@embark-status): Fix support for IPFS
Previously, `embark-status` was meant to support IPFS but was experiencing CORS issues that would not allow successful API/gateway requests when being served to non-localhost URLs.

This PR includes a change for using including the port in the URL added to IPFS CORS, and moves the CORS update call to after the webserver is loaded.

For this funcationality to work correctly, https://github.com/embark-framework/embark/pull/1139 is needed.